requirements.txt
flask>=2.3.0
Flask-SQLAlchemy>=3.0.0
flask-login>=0.6.0
python-dotenv>=1.0.0
Werkzeug>=2.3.0
email-validator>=2.0.0
pytest>=8.4.2